==> 2. Creating Basket Data.ipynb <==
# 2. Creating Basket Data + DTM Data

The purpose of this notebook is to generate basket data, which is good for clusering and association roles mining. The desired shape of the table has each row is a document, and the columns are the individual words. Finally, the output will be in a csv format.

The data used in this notebook are from the Count Vectorizer. This is used so that the values in the CountVectorizer can be converted to binary, and then the row can be cleaned.

## 1. Environment Creation

### 1.1 Library Import

import pandas as pd
import os
from tqdm.notebook import tqdm
import csv

### 1.2 Data Import

bill_information = pd.read_csv(r"C:\Users\natal\OneDrive\university\info 5653\data\Bills Lemmed- Count Vectorizer.csv")

bill_information.head(2)

news = pd.read_csv(r"C:\Users\natal\OneDrive\university\info 5653\data\News Articles Lemmed- Count Vectorizer.csv")

news.head(2)

party_platform = pd.read_csv(r"C:\Users\natal\OneDrive\university\info 5653\data\Party Platform Lemmed- Count Vectorizer.csv")

party_platform.head(2)

### 1.2.2 Light Data Cleaning

The data utilized in this notebook has already undergone extensive cleaning processes, which may be viewed on [my project page](https://nataliermcastro.github.io/projects/2025/01/14/political-stances-data.html). In this section, the labels will be removed and the unnamed column will be dropped.

bill_information.drop(columns=['Unnamed: 0','Sponser State','Bill Type','Sponser Affiliation','Committees'],inplace=True)

bill_information.drop(columns=['helvetica','noto','sego','neue','vh','html','webkit','emoji','blinkmacsystemfont','arial','roboto','ui','serif',
                              'column','font','pad','width','auto','left','height'],inplace=True)

news.drop(columns=['Unnamed: 0','Party','publisher'],inplace=True)

party_platform.drop(columns=['Unnamed: 0','Party'],inplace=True)

## 2. Creating Basket Data

### 2.1 Assembling the Transactions

def transaction_creator(index,basket):
    transcation_dictionary = {key:val for key, val in basket[index].items() if val != 0.0}
    items = list(transcation_dictionary.keys())
    return (items)

news_basket = news.to_dict(orient='records')
bills_basket = bill_information.to_dict(orient='records')
party_basket = party_platform.to_dict(orient='records')

baskets = [news_basket,bills_basket,party_basket]

transactions = []

for current_basket in baskets:
    for index in tqdm(range(0,len(current_basket)),desc='🛒🐛... | inching through the store'):
        transactions.append(transaction_creator(index,current_basket))
        

## And checking to make sure it worked...
transactions[0]

### 2.2 Saving the Basket Data

''' WRITING TO A CSV '''

with open('Basket Data.csv', 'w', newline='\n') as csvfile:
    writer = csv.writer(csvfile)
    writer.writerows(transactions)

## 3. Creating Document Term Matrix Data

dtm_df = pd.concat([bill_information,news, party_platform])

dtm_df.head()

dtm_df = dtm_df.fillna(0)

dtm_df.reset_index(inplace=True,drop=True)

dtm_df.head(1)

dtm_df.to_csv("Document Term Matrix.csv")
